requirements.txt
pdf2image
paddleocr
Pillow
opencv-python
numpy
easyocr
tk
